"Implemented rule name editability and TableView initialization improvement"
In the RuleController, the rule name field was set to be non-editable to prevent the user from changing the original rule name during the editing process.

MainController now initializes TableView with proper cell factories and event handlers to handle rule edition. The rule list is now loaded from the backup file to an ObservableList rules, which is set as the TableView items.

In the main-view.fxml file, TableView was set to be editable to allow users to edit the rule's names from the TableView and TableColumn 'indexClm' was set to be non-editable.