Add CompositeOrTrigger class and related tests
This commit introduces a new class CompositeOrTrigger. This class extends CompositeTrigger and overrides the evaluate() method. The code also includes a comprehensive test for this new class. There's also a tweak made in the CompositeAndTrigger class to streamline the evaluate() method. The unnecessary getTYPE() override in NotTriggerDecoratorTest has been removed.