Refactor view packages and add negation option
Moved Action and Trigger view classes into dedicated subpackages to make the project structure more intuitive. Added negation functionality to triggers, providing the user option to invert the evaluation of individual triggers. This introduced a checkbox in the `ElementaryTriggerView` and also a toString method in `NotTriggerDecorator` to display proper descriptions.